workshop/requirements.txt
scipy
scikit-learn
datasets
tensorflow-cpu
transformers
tf-keras
Pillow
matplotlib